+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorCustomValidatorTest.kt
@@ -46,7 +46,7 @@ class CodeGeneratorCustomValidatorTest {
         parser.customValidationHandler = { key, uri, location, value ->
             when (key) {
                 "x-test" -> {
-                    if (value is JSONString && value.get() == "not-empty")
+                    if (value is JSONString && value.value == "not-empty")
                         StringValidator(uri, location, StringValidator.ValidationType.MIN_LENGTH, 1)
                     else
                         throw RuntimeException("Unknown type")
@@ -69,7 +69,7 @@ class CodeGeneratorCustomValidatorTest {
         parser.customValidationHandler = { key, uri, location, value ->
             when (key) {
                 "x-test" -> {
-                    if (value is JSONString && value.get() == "not-empty")
+                    if (value is JSONString && value.value == "not-empty")
                         StringValidator(uri, location, StringValidator.ValidationType.MIN_LENGTH, 1)
                     else
                         throw RuntimeException("Unknown type")